requirements.txt
transformers==4.33.1
accelerate==0.23.0
diffusers==0.22.0
gradio==4.8.0
pyyaml
pygame
opencv-python
info-nce-pytorch
kornia
segmentation-models-pytorch==0.3.3
tqdm
torch==1.13.1+cu117 
torchvision==0.14.1+cu117 
torchaudio==0.13.1 --extra-index-url https://download.pytorch.org/whl/cu117
Pillow 
numpy 
tensorboard
